refactor: rename context-selection to report-parameter-selector
