+++ src/components/report-parameter-selector/PeriodSelector.js
