+++ src/pages/annual_report/AnnualReport.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import { ReportData } from '../../components/annual-report/report-data/ReportData.js'
-import { ReportParameterSelector } from '../../components/context-selection/index.js'
 import MenuBar from '../../components/menu-bar/MenuBar.js'
+import { ReportParameterSelector } from '../../components/report-parameter-selector/index.js'
 
 const Report = () => {
     const [reportParameters, setReportParameters] = useState({})
